% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.sol}
\alias{data.sol}
\title{Data.Sol}
\format{
An object of class \code{data.frame} with 3153 rows and 460 columns.
}
\usage{
data.sol
}
\description{
Average shortwave radiation on horizontal surfaces in W/m², listed by geographical coordinates for Germany = latitude and longitude distances 0.02°.
}
\keyword{datasets}
