% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list.station.ta}
\alias{list.station.ta}
\title{List.Station.TA}
\format{
An object of class \code{data.frame} with 843 rows and 10 columns.
}
\usage{
list.station.ta
}
\description{
Provides the geographical coordinates for all DWD stations.
}
\keyword{datasets}
