% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab.estim.sol.orient}
\alias{tab.estim.sol.orient}
\title{Tab.Estim.Sol.Orient}
\format{
An object of class \code{data.frame} with 33 rows and 13 columns.
}
\usage{
tab.estim.sol.orient
}
\description{
Includes parameters for estimating the solar radiation for vertical and inclined surfaces on the basis of horizontal radiation data.
}
\keyword{datasets}
