% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab.stationmapping}
\alias{tab.stationmapping}
\title{Tab.StationMapping}
\format{
An object of class \code{data.frame} with 16336 rows and 17 columns.
}
\usage{
tab.stationmapping
}
\description{
Assigns to all German postal codes the geographical coordinates and the three closest DWD weather stations.
}
\keyword{datasets}
