% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.ta.hd}
\alias{data.ta.hd}
\title{Data.TA.HD}
\format{
An object of class \code{data.frame} with 11651 rows and 460 columns.
}
\usage{
data.ta.hd
}
\description{
Per DWD climate station: average external air temperature in °C; for specific base temperatures: heating and cooling days, external air temperature during heating and during cooling days.
}
\keyword{datasets}
